extract syslogd server
