--- internal/logd/README.md
@@ -1,1 +1,0 @@
-Provides a UDP-based Syslog server.
